requirements.txt
gdown
scipy
pillow
pandas
opencv-python
torch
torchvision
wandb
# segmentation dependencies
pip install pycocotools
pip install segmentation-models-pytorch